tab is now bound to dabbrev-expand. This allows users to complete projects and tags using tab.
Try the following @cc<TAB>. You should get the first tag starting
with @cc (if more than one is present, pressing TAB again rotates
among all possible choices).

The completion works for any word, not necessarily tags and projects
(but it will probably be more useful there).

To insert a tab in the file use "C-q TAB" (but maybe the need for
tabs in the todo.txt file is low)